requirements.txt
yfinance
pandas
numpy
streamlit
beautifulsoup4
requests
google-genai
python-dotenv
plotly
scikit-learn
joblib
pyts
matplotlib
opencv-python-headless
tqdm
xgboost
prophet
holidays
cmdstanpy
feedparser

gymnasium
stable-baselines3
